Move: Position 관련 타입 이동
